+++ Data/Entities/UserDetails.cs
@@ -13,9 +13,9 @@ namespace FitnessPortal.Data.Entities
         public string LastName { get; set; }
         public string? Gender { get; set; }
         public int? Age { get; set; }
-        public float? Weight { get; set; }
-        public float? Height { get; set; }
-        public float? BMI { get; set; }
+        public double? Weight { get; set; }
+        public double? Height { get; set; }
+        public double? BMI { get; set; }
         public int? KCalGoal { get; set; }
         [ForeignKey("UserID")]
         public Guid UserID { get; set; }
